+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Actually add Summon Lesser Servitor...
 - Fixed Dinosaur Form (Brontosaurus) scaling.
 - Added automatic effects to Inspire Courage, Inspire Defense, Bless, and Protective Ward, along with PF2e Workbench'es variant effects if they exist.
+- Fixed a bunch of bugs relating to summoning filter args.
 
 # Version 2.7.2
 - Removed soundfxlibrary dependency. Animations with sounds will not play the sound if it is missing.